+++ rusqlite/Changelog.md
@@ -1,5 +1,6 @@
 # Version UPCOMING (TBD)
 
+* Renamed `SqliteConnection` to `Connection`. The old name remains as a typealias.
 * Adds a variety of `..._named` methods for executing queries using named placeholder parameters.
 * Adds `backup` feature that exposes SQLite's online backup API.
 * Adds `functions` feature that allows user-defined scalar functions to be added to